refactor(): move ajax routes to config node
