+++ src/nodes/tahoma.ts
@@ -1,5 +1,4 @@
 import { Red } from 'node-red';
-import * as fs from 'fs';
 import { ICommand } from '../interfaces/command';
 import { SomfyApi } from '../core/somfy-api';
 import { ICommandExecutionResponse } from '../interfaces/command-execution-response';
@@ -132,28 +131,5 @@ export = (RED: Red) => {
                     });
                 });
         });
-    });
-
-    RED.httpAdmin.get('/somfy/callback', (request, response) => {
-        const CALLBACK_BODY = fs.readFileSync(__dirname + '/somfy-callback.html');
-        response.header('Content-Type', 'text/html');
-        response.write(CALLBACK_BODY.toString());
-        response.send();
-    });
-
-    RED.httpAdmin.get('/somfy/:account/sites', function (req, res) {
-        const configNode = RED.nodes.getNode(req.params.account) as any;
-        const somfyApiClient = new SomfyApi(RED, configNode.context, req.params.account);
-
-        somfyApiClient.getSites()
-            .then((sites: any) => res.json(sites));
-    });
-
-    RED.httpAdmin.get('/somfy/:account/site/:siteid/devices', function (req, res) {
-        const configNode = RED.nodes.getNode(req.params.account) as any;
-        const somfyApiClient = new SomfyApi(RED, configNode.context, req.params.account);
-
-        somfyApiClient.getDevicesForSite(req.params.siteid)
-            .then((devices: any) => res.json(devices));
     });
 };
